# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Igloo",
    "description": "The Igloo control plane API.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/v1/changes/{id}": {
      "get": {
        "tags": [
          "changes"
        ],
        "summary": "Gets a change.",
        "operationId": "getChange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/changes/{id}/approve": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Approves a change's latest revision as the caller, who must be a human.",
        "operationId": "approveChange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/changes/{id}/close": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Closes a change without merging.",
        "operationId": "closeChange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/changes/{id}/comments": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Comments on a revision of a change, the latest by default, optionally on a line of a file.",
        "operationId": "commentOnChange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/changes/{id}/merge": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Merges a change: its latest revision's checks must have passed, it must be based on the\ntarget branch's head, and changes to protected paths need a human approval. Igloo pushes\nthe revision to the target branch as one squashed commit naming the change.",
        "operationId": "mergeChange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/changes/{id}/request-changes": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Asks for another revision, carrying the comments made since the previous request. A change\nmade by a task sends them to its agent as its next turn.",
        "operationId": "requestChanges",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/changes/{id}/revisions": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Records the source branch's current head as the next revision; an unchanged head records\nnothing.",
        "operationId": "reviseChange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/repos/{id}/changes": {
      "get": {
        "tags": [
          "changes"
        ],
        "summary": "Lists a repository's changes, oldest first unless `order=newest`, optionally only those at\nthe given phases. Open changes carry their merge readiness.",
        "operationId": "listRepoChanges",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phase",
            "in": "query",
            "description": "Only changes at this phase; repeatable",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ChangePhase"
              }
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "`newest` lists the newest first",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ListOrder"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChangeResource"
                  }
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Opens a change proposing a branch pushed to the forge; its head is the first revision.",
        "operationId": "openChange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApprovalNeed": {
        "type": "object",
        "description": "The approval a change needs.",
        "required": [
          "state"
        ],
        "properties": {
          "protected_paths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The protected paths the change touches; empty when `state` is `not_required`."
          },
          "state": {
            "$ref": "#/components/schemas/ApprovalState",
            "description": "Whether it is needed and given."
          }
        }
      },
      "ApprovalResource": {
        "type": "object",
        "description": "A human's approval of a revision.",
        "required": [
          "revision",
          "by",
          "at"
        ],
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "description": "When."
          },
          "by": {
            "type": "string",
            "description": "Who approved it (`usr_...`)."
          },
          "revision": {
            "type": "integer",
            "format": "int32",
            "description": "The revision approved.",
            "minimum": 0
          }
        }
      },
...
      "ApproveRequest": {
        "type": "object",
        "description": "Approves a change's latest revision.",
        "properties": {
          "revision": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The revision approved; it must be the latest. The latest when omitted.",
            "minimum": 0
          }
        }
      },
      "ChangePhase": {
        "type": "string",
        "description": "Where a change is.",
        "enum": [
          "open",
          "merged",
          "closed"
        ]
      },
      "ChangeResource": {
        "type": "object",
        "description": "A change.",
        "required": [
          "id",
          "repo",
          "source_branch",
          "target_branch",
          "title",
          "phase",
          "revisions",
          "approvals"
        ],
        "properties": {
          "approvals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApprovalResource"
            },
            "description": "Human approvals, oldest first."
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResource"
            },
            "description": "Review comments, oldest first."
          },
          "id": {
            "type": "string",
            "description": "Its id (`chg_...`)."
          },
          "merged_commit": {
            "type": [
              "string",
              "null"
            ],
            "description": "The target branch's head after the merge, once merged."
          },
          "phase": {
            "$ref": "#/components/schemas/ChangePhase",
            "description": "Where it is."
          },
          "readiness": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MergeReadiness",
                "description": "Merge readiness; present only while the change is open."
              },
              {
                "type": "null"
              }
            ]
          },
          "repo": {
            "type": "string",
            "description": "The repository."
          },
          "revisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RevisionResource"
            },
            "description": "Every revision, oldest first."
          },
          "source_branch": {
            "type": "string",
            "description": "The branch proposed."
          },
          "target_branch": {
            "type": "string",
            "description": "The branch it merges into."
          },
          "title": {
            "type": "string",
            "description": "What it does."
          }
        }
      },
...
      "ChecksState": {
        "type": "string",
        "description": "Where the checks of a change's latest revision are.",
        "enum": [
          "passed",
          "failed",
          "running",
          "missing"
        ]
      },
      "CommentRequest": {
        "type": "object",
        "description": "Comments on a change.",
        "required": [
          "body"
        ],
        "properties": {
          "body": {
            "type": "string",
            "description": "What it says; 1 to 10 000 characters."
          },
          "line": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The line of `path` it is about, from 1.",
            "minimum": 0
          },
          "path": {
            "type": [
              "string",
              "null"
            ],
            "description": "The file it is about, relative to the repository root."
          },
          "revision": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The revision; the latest when omitted.",
            "minimum": 0
          }
        }
      },
      "CommentResource": {
        "type": "object",
        "description": "A reviewer's comment on a revision.",
        "required": [
          "revision",
          "author",
          "body",
          "at"
        ],
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "description": "When."
          },
          "author": {
            "type": "string",
            "description": "Who wrote it: a user (`usr_...`), an agent (`agt_...`) or `system`."
          },
          "body": {
            "type": "string",
            "description": "What it says."
          },
          "line": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The line of `path` it is about, from 1.",
            "minimum": 0
          },
          "path": {
            "type": [
              "string",
              "null"
            ],
            "description": "The file it is about."
          },
          "revision": {
            "type": "integer",
            "format": "int32",
            "description": "The revision commented on.",
            "minimum": 0
          }
        }
      },
...
      "FieldError": {
        "type": "object",
        "description": "One invalid field of a request.",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string",
            "description": "The field path, such as `\"limits.millicpus\"`."
          },
          "message": {
            "type": "string",
            "description": "Why it is invalid."
          }
        }
      },
...
      "ListOrder": {
        "type": "string",
        "description": "The order of a list. Lists are oldest first unless asked otherwise.",
        "enum": [
          "oldest",
          "newest"
        ]
      },
...
      "MergeReadiness": {
        "type": "object",
        "description": "What stands between an open change and its merge, by the rules merging applies. Merging\nstays authoritative: the target may move after this was computed.",
        "required": [
          "checks",
          "approval",
          "fast_forward"
        ],
        "properties": {
          "approval": {
            "$ref": "#/components/schemas/ApprovalNeed",
            "description": "The approval of the latest revision."
          },
          "checks": {
            "$ref": "#/components/schemas/ChecksState",
            "description": "The checks of the latest revision."
          },
          "fast_forward": {
            "type": "boolean",
            "description": "Whether the latest revision's base is the target branch's head, so it merges as is."
          }
        }
      },
...
      "OpenChangeRequest": {
        "type": "object",
        "description": "Opens a change for a branch pushed to the forge.",
        "required": [
          "branch",
          "title"
        ],
        "properties": {
          "branch": {
            "type": "string",
            "description": "The branch proposed."
          },
          "title": {
            "type": "string",
            "description": "What the change does, 1 to 200 characters."
          }
        }
      },
      "Problem": {
        "type": "object",
        "description": "An error response. `code` is stable and machine-readable; `title` and `detail` are for\nhumans.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "The stable error code, such as `\"sandbox.not_found\"`."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "What went wrong in this occurrence."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "Every invalid field, for validation problems."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "A short summary of the problem type."
          },
          "type": {
            "type": "string",
            "description": "A URI identifying the problem type."
          }
        }
      },
...
      "RevisionResource": {
        "type": "object",
        "description": "One revision of a change.",
        "required": [
          "number",
          "head",
          "base",
          "created_at"
        ],
        "properties": {
          "base": {
            "type": "string",
            "description": "Where it forked from the target branch."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When it was recorded."
          },
          "head": {
            "type": "string",
            "description": "The source branch's head."
          },
          "number": {
            "type": "integer",
            "format": "int32",
            "description": "Its position, from 1.",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}